backend/requirements.txt
flask
pyyaml
taipy
